Allow `,` in parents of template
New syntax

    A extends B, C { ... }

instead of

    A extends B with C { ... }

The old syntax is still allowed, but it's planned to phase it out together
with `with` as a type operator.

Also: disallow

    A extends { ... }

(make it a migration warning under -language:Scala2)

While doing these changes, applied some refactorings to handle templates
in the parser which hopefully make the code clearer.